+++ pennylane_lightning/core/src/simulators/lightning_gpu/utils/cuStateVecError.hpp
@@ -1,0 +1,105 @@
+// Copyright 2022-2023 Xanadu Quantum Technologies Inc.
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+
+//     http://www.apache.org/licenses/LICENSE-2.0
+
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Adapted from JET: https://github.com/XanaduAI/jet.git
+// and from Lightning: https://github.com/PennylaneAI/pennylane-lightning.git
+/**
+ * @file cuStateVecError.hpp
+ */
+
+#pragma once
+#include <string>
+
+#include <custatevec.h>
+
+#include "Error.hpp"
+#include "Util.hpp"
+// LCOV_EXCL_START
+namespace {
+using namespace Pennylane::Util;
+}
+
+#ifndef CUDA_UNSAFE
+
+/**
+ * @brief Macro that throws Exception from cuStateVec failure error codes.
+ *
+ * @param err cuStateVec function error-code.
+ */
+#define PL_CUSTATEVEC_IS_SUCCESS(err)                                          \
+    PL_ABORT_IF_NOT(                                                           \
+        err == CUSTATEVEC_STATUS_SUCCESS,                                      \
+        Pennylane::LightningGPU::Util::GetCuStateVecErrorString(err).c_str())
+
+#else
+#define PL_CUSTATEVEC_IS_SUCCESS(err)                                          \
+    { static_cast<void>(err); }
+#endif
+
+namespace Pennylane::LightningGPU::Util {
+static const std::string
+GetCuStateVecErrorString(const custatevecStatus_t &err) {
+    std::string result;
+    switch (err) {
+    case CUSTATEVEC_STATUS_SUCCESS:
+        result = "No errors";
+        break;
+    case CUSTATEVEC_STATUS_NOT_INITIALIZED:
+        result = "custatevec not initialized";
+        break;
+    case CUSTATEVEC_STATUS_ALLOC_FAILED:
+        result = "custatevec memory allocation failed";
+        break;
+    case CUSTATEVEC_STATUS_INVALID_VALUE:
+        result = "custatevec invalid value";
+        break;
+    case CUSTATEVEC_STATUS_ARCH_MISMATCH:
+        result = "custatevec CUDA device architecture mismatch";
+        break;
+    case CUSTATEVEC_STATUS_EXECUTION_FAILED:
+        result = "custatevec execution failed";
+        break;
+    case CUSTATEVEC_STATUS_INTERNAL_ERROR:
+        result = "custatevec internal error";
+        break;
+    case CUSTATEVEC_STATUS_NOT_SUPPORTED:
+        result = "custatevec unsupported operation/device";
+        break;
+    case CUSTATEVEC_STATUS_INSUFFICIENT_WORKSPACE:
+        result =
+            "custatevec insufficient memory for gate-application workspace";
+        break;
+    case CUSTATEVEC_STATUS_SAMPLER_NOT_PREPROCESSED:
+        result = "custatevec sampler not preprocessed";
+        break;
+    case CUSTATEVEC_STATUS_NO_DEVICE_ALLOCATOR:
+        result = "custatevec no device allocator";
+        break;
+    case CUSTATEVEC_STATUS_DEVICE_ALLOCATOR_ERROR:
+        result = "custatevec device allocator error";
+        break;
+    case CUSTATEVEC_STATUS_COMMUNICATOR_ERROR:
+        result = "custatevec communicator failure";
+        break;
+    case CUSTATEVEC_STATUS_LOADING_LIBRARY_FAILED:
+        result = "custatevec dynamic library load failure";
+        break;
+    default:
+        result =
+            "custatevec status not found. Error code=" + std::to_string(err);
+    }
+    return result;
+}
+} // namespace Pennylane::LightningGPU::Util
+  // LCOV_EXCL_STOP
