+++ pennylane_lightning/core/src/utils/config.h
@@ -14,10 +14,11 @@
 
 /**
  * @file
- * Record the path to scipy.libs at compile time.
+ * Config file for the path to scipy.libs at compile time.
  */
 
 #ifndef CONFIG_H
 #define CONFIG_H
-#define SCIPY_LIBS_PATH ""
+#define SCIPY_LIBS_PATH                                                        \
+    "/home/shuli/env/lib/python3.10/site-packages/scipy.libs"
 #endif
